Simplify process for consolidating status
