+++ pkg/apis/hco/v1alpha1/hyperconverged_types.go
@@ -9,6 +9,7 @@ import (
 
 // EDIT THIS FILE!  THIS IS SCAFFOLDING FOR YOU TO OWN!
 // NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
+// Important: Run "operator-sdk generate k8s" to regenerate code after modifying this file
 
 // HyperConvergedSpec defines the desired state of HyperConverged
 // +k8s:openapi-gen=true
@@ -21,15 +22,22 @@ type HyperConvergedSpec struct {
 // HyperConvergedStatus defines the observed state of HyperConverged
 // +k8s:openapi-gen=true
 type HyperConvergedStatus struct {
-	// INSERT ADDITIONAL STATUS FIELD - define observed state of cluster
-	// Important: Run "operator-sdk generate k8s" to regenerate code after modifying this file
-	// Add custom validation using kubebuilder tags: https://book.kubebuilder.io/beyond_basics/generating_crd.html
-
-	// Conditions is a list of conditions related to operator reconciliation
+	// Conditions describes the state of the HyperConverged resource.
+	// +patchMergeKey=type
+	// +patchStrategy=merge
+	// +optional
 	Conditions []conditionsv1.Condition `json:"conditions,omitempty"  patchStrategy:"merge" patchMergeKey:"type"`
-	// RelatedObjects is a list of objects that are "interesting" or related to this operator.
+
+	// RelatedObjects is a list of objects created and maintained by this
+	// operator. Object references will be added to this list after they have
+	// been created AND found in the cluster.
+	// +optional
 	RelatedObjects []corev1.ObjectReference `json:"relatedObjects,omitempty"`
 }
+
+// ConditionReconcileComplete communicates the status of the HyperConverged resource's
+// reconcile functionality. Basically, is the Reconcile function running to completion.
+const ConditionReconcileComplete conditionsv1.ConditionType = "ReconcileComplete"
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
